refactor: Remove unneeded `disabled` property
Removed the internal usage of `disabled` attribute which was initially used to determine the state of 'clickable'. Instead, directly using 'clickable', a property of the card class, for both the primary action class and ripple effect. This led to more succinct and readable code, reducing unnecessary state checks and updates. It also eliminated internal TODO comments regarding this issue.